CheckOCRScreeningStatus modul: bugfixes

+++ AktBob.CheckOCRScreeningStatus/ModuleServices.cs
@@ -13,7 +13,7 @@ public static class ModuleServices
     {
         // Guard against missing Podio configuration
         var podioAppId = Guard.Against.Null(configuration.GetValue<int>("Podio:AppId"));
-        var podioFields = Guard.Against.Null(Guard.Against.NullOrEmpty(configuration.GetSection("Podio:Fields").GetChildren().ToDictionary(x => long.Parse(x.Key), x => x.Get<PodioField>())));
+        var podioFields = Guard.Against.Null(Guard.Against.NullOrEmpty(configuration.GetSection("Podio:Fields").GetChildren().ToDictionary(x => int.Parse(x.Key), x => x.Get<PodioField>())));
         var podioFieldFilArkivCaseId = Guard.Against.Null(podioFields.FirstOrDefault(x => x.Value.AppId == podioAppId && x.Value.Label == "FilArkivCaseId"));
         Guard.Against.Null(podioFieldFilArkivCaseId.Value);
